cyber-farmers-backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
diff-match-patch==20241021
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cloudinary-storage==0.3.0
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.1
django-jazzmin==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Faker==33.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.2
tablib==3.7.0
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
